Cleanup and update hero stats menu for variable res and higher stat limits
Higher stat limits meant more space needed, and required a change to the
way that the graph is drawn.

However, it is simply impossible to fit the hero stats menu into 320x200 if the
max level isn't 99 (causing a 3rd menu to appear). I tried my best to limit
overlap, but there's still lots.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9332 7d344553-34f0-0310-a9b1-970ce8f1c3a2