spriteedit: change a number of mouse click checks into button-down or -release checks
This fixes several problems where clicking would enter a menu and then releasing
would cause a misclick in that menu

However, I left all the tool buttons (except import/export) as the old behaviour
for now. Not sure about it. Might become a mess if I try to change anything,
since it is a bit complicated, with some buttons like Rotate being special

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9445 7d344553-34f0-0310-a9b1-970ce8f1c3a2