Fix gfx_sdl's io_keybits not reporting new-keypresses correctly - it was using key repeat.
I removed a previous workaround for keys getting stuck when spawning a console
program, which didn't actually work.

However, I've discovered that on Windows a second type of key-repeat happens
due to the OS, where keyup-keydown event pairs get sent, and it doesn't
seem to be possible to filter them out 100% of the time because sometimes
the events are delivered on different ticks!
gfx_fb shows the same problem.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9479 7d344553-34f0-0310-a9b1-970ce8f1c3a2