Replace old hash_file() with file_hash_SHA1() and file_hash64() functions, and other SHA1 wrappers
file_hash64 returns the first 64 bits of the SHA1 of the file.
So instead of using a bad 32 bit hash, which could have caused collisions
concievably resulting in files not getting reloaded, we use a good 64 bit hash.
(Yes, I know it's not cryptographically secure, but we won't be using this for
security)

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9529 7d344553-34f0-0310-a9b1-970ce8f1c3a2