Fix crash when using F9 Test Game from inside Import Scripts, and other places waitforanykey or fadeto are called
Use of the last_setvispage variable was dangerous, so I added a getvispage()
function to return the current visible page, which is sometimes different from
vpage.
We ought to either ensure that vpage (or a view onto it) is use everywhere,
or use getvispage everywhere that draws something on top of an existing screen.
Not yet sure which approach to use.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9539 7d344553-34f0-0310-a9b1-970ce8f1c3a2