mapedit: Allow panning the map by dragging with the right mouse button.
My preference would be for panning to just be dragging the map around, possibly
contining to scroll if you drag to the window edge, but instead this is the
velocity sort of panning. Might still chance it later... but by then everyone
will be used to it!

I allow the cursor to go offscreen, because it looks back to have it trailing
behind as the map pans.

Map layer skewing now is done by holding CTRL while dragging with the right
mouse button. Hmm, I guess I could probably have gone with CTRL+left mouse.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9757 7d344553-34f0-0310-a9b1-970ce8f1c3a2