mapedit: Mouse support for NPC placement mode.
This mode acts differently from the others, because left clicking places/deletes
and NPC but doesn't move the cursor. Otherwise you would get an ugly overlap.
I guess it would be possible to make the cursor hover underneath the mouse;
I think the clone tool should probably do that too.

Currently, it's not possible to move the cursor onto an NPC with the mouse,
so you can't see its avoidance zones.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9758 7d344553-34f0-0310-a9b1-970ce8f1c3a2