Add findrgb(), as using UI cols in Custom is madness. Change mapedit position finder colors
Yes, these colors are now given as RGB triples instead of uicol constants.
(It was too hard to see the zone minimap over the screen position rect)

I still think we should have a second UI cols menu with editor colours,
basically for theming, but it would be too much trouble to put everything there.
But we should definitely stop using in-game UI cols in Custom.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9796 7d344553-34f0-0310-a9b1-970ce8f1c3a2