Fix weird tile animation pattern editor behaviour, and document what happens if there's no "end of animation"
It turns out that at least three games do actually make use of the secret ability
to loop through the whole tileset this way: Timpoline, Super Walrus Chef, and
H-Metroid (in Barnabus's All The Games). And I used it in an unreleased game.

Now you can't edit animation ops past the end of the animation anymore. And if
there's no "end of animation", then "repeat" is displayed after the end instead.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9807 7d344553-34f0-0310-a9b1-970ce8f1c3a2