(Mostly) fix wrapped text slice height not updating when the width changed.
It's still not updated until DrawTextSlice is called, meaning there's a one
tick delay. I think ideally it would instantly update when you call
setslicewidth, like it already does when you call setslicetext.

This bug was hidden by the fact that the slice editor manually called
UpdateTextSlice.

Turns out that this bug was already documented in the source, but I hadn't
done anything about it.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9821 7d344553-34f0-0310-a9b1-970ce8f1c3a2