Convert show_help to use a scroll slice and new-style text slice
Instead of using first_line or line_limit (which aren't supported by
wrapprint-based text slices), the text slice is scrolled up and down.
However, still switches the text slice back to old-style wrapping when
editing, so that the cursor is drawn correctly.

Incidentally this results in more CPU usage when editing, and less when viewing.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9836 7d344553-34f0-0310-a9b1-970ce8f1c3a2