Fix major memory leak while recording a gif. And now always preserve vpage
Every page fed to setvispage was duplicated and leaked.

It seems that I was prematurely optimising in trying to avoid perserving vpage;
actually the cost of the memcpy is pretty minor. I had timed it before and
found it was a major cost when running at 100's of fps, but that must have been
before I implemented frameskipping.

So this also prevents the possibility of overlays becoming permanently plastered
onto the screen (when making a copy of vpage)

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9868 7d344553-34f0-0310-a9b1-970ce8f1c3a2