Fix characters above 127 in text not being copy-pastable
Now encodes text as UTF8 when putting it on the clipboard, as I was meant to.
Added latin1_to_utf8().

I haven't fixed the other clipboard problem on Windows yet, though.

Also, utf8_to_latin1() now returns the string instead of "[CORRUPTED]" if it's
not valid UTF8.
Strangely, copy pasting on Linux, but not Windows, resulted in getting back
[CORRUPTED].

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9871 7d344553-34f0-0310-a9b1-970ce8f1c3a2