scons: try to fix link errors on linux distros which default to PIE on non-x86, when using linkgcc=1
Also try to handle differences in the spelling of the -no-pie option in different
gcc and clang versions. Tried to merge all the PIE stuff to one big block.

Not actually tested with either old gcc/clang, nor on a distro defaulting to PIE.
Mac and windows builds should be unaffected by this change.
linkgcc=0 already worked fine.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9906 7d344553-34f0-0310-a9b1-970ce8f1c3a2