Fix editing of attack elemental fail conditions, recently broken in r8948
Any edits to the condition thresholds in the percent_cond_editor submenu were
erased when you left the menu. Editing them out of the menu also suffered
brokenness, and the conditions got overwritten if one was the active menu item
when you switched between attacks.
r8948 was completely wrong. The captions are only meant to reset when you load
an attack.

percent_grabber now resets repr (the caption) itself if it detects it's incorrect
rather than overwriting data with garbage

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9913 7d344553-34f0-0310-a9b1-970ce8f1c3a2