Support loading progressive JPEGs by running them through the jpegtran tool
This small utility program is part of jpeglib-progs/jpeglib-turbo-progs so is
typically installed on Linux machines.
On Windows, it will be downloaded when needed. You're only asked to download it
once per time Custom is launched.
This utility also has some other uses, like supporting jpegs with arithmetic
coding, but those are very rare.

I patched ujpeg so that it retrieves the size and so forth of progressive jpegs
although it can't read them.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9925 7d344553-34f0-0310-a9b1-970ce8f1c3a2