Fix two bugs involving textbox 'after' links, one quite nasty.
In the textbox editor, pressing Ctrl+Left/Right to link to another box appeared
to work but actually the "After" tag condition didn't get set to "Always",
although it claimed it was. It didn't actually get set to "Always" until you
entered a submenu. The textbox editor doesn't save when you exit (instead, its
meant to save when data is edited), so it's possible that what appears to be the
state of the textbox never makes it to disk.
Reported by Taoki.

The in-game F1 textbox debug screen also showed "ALWAYS" instead of "NEVER"
(this only affected the After box/script option; tags aren't shown for other
conditionals), which made this bug even more confusing.

The bug was that these lines of code were in the wrong order:
       SaveTextBox box, st.id
       update_textbox_editor_main_menu box, menu()
but to make this less confusing I instead split out "textbox_set_after_textbox"

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9931 7d344553-34f0-0310-a9b1-970ce8f1c3a2