Extend Frame <-> Node saving/loading to support Frame arrays, not just isolated Frames
The nodes are now named "frameset" instead of "frame".
For now, all Frames in the array must be the same size (we don't have any
function to allocate anything else anyway).
I had been planning a condensed format for isolated frames but it would only
save a single extra "frame" child, so not worth the trouble.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9995 7d344553-34f0-0310-a9b1-970ce8f1c3a2