When upgrading, unconditionally create defpal#.bin if they don't already exist.
This fixes the "Palette: default" option not working anywhere in old games until
you entered and exited the spriteset editor for that sprite type (either the old
or new editors). Looked really bad in the new spriteset editor, because it uses
sprite slices with default palettes.

Surprisingly, test.rpg is still missing the default palettes for most sprite
types!

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10001 7d344553-34f0-0310-a9b1-970ce8f1c3a2