scons: Reduce game/custom.exe size by 1MB by compiling with --gc-sections
Measured on a gengcc=1 debug=0 build with MXE (mingw-w64).
Mainly, this cuts out most of libstdc++, which we statically link. So the effect
isn't dramatic on other platforms.

I now measure game.exe as 3MB and custom.exe as 3.5MB. Of that, 600KB is
symbol information, unfortunately a large amount of that is from libstdc++,
because ld doesn't delete symbols for GC'd sections.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10025 7d344553-34f0-0310-a9b1-970ce8f1c3a2