ON ERROR GOTO QB_ERROR_HANDLER
Didn't think I'd be re-adding QBasic code to the engine...
If compiled with -exx, when FB throws an error you now get a graphical
fatalerror() screen (assuming it doesn't crash in the process), and of course
it's logged to g/c_debug.txt. Much better than the error message ending up
in the bit bucket on Windows and the program instantly closing!

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10039 7d344553-34f0-0310-a9b1-970ce8f1c3a2