Rename the Caption menu item type to Label.
The double use of Caption for both an item type, and the caption of a item,
as well as the fact that these are unrelated to attack captions, and the normal
meaning of "caption" for that matter, seemed like a poor word choice.

Actually, they were already called labels in the plotdict.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10183 7d344553-34f0-0310-a9b1-970ce8f1c3a2