Also make the Margins special menu item un-activatable, for consistency with Music and Sound Volume
Of course this is a back-compat-breaking change, if someone actually did
something so strange as to set a tag when attempting to use the Margins option.
I scanned 1613 games and found one game (Macabre, funnily enough) which had
Margins set to Close menu when activated. Which is safe to ignore.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10185 7d344553-34f0-0310-a9b1-970ce8f1c3a2