Add push_and_reset_gfxio_state/pop_gfxio_state and make use of it
This enables the mouse, resets the palette, and makes the window resizable,
and then undoes that.
As a result, you can now resize the window in a number of debug menus including
the show_help, multichoice, live_preview_menu, etc.
Many of these menus don't actually resize to make use of the space if you
resize the window

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10273 7d344553-34f0-0310-a9b1-970ce8f1c3a2