Enable mouse controls in Game's rpg browser.
I was going to make the browser window resizable too, but it turns out
this triggers an annoying SDL bug on Linux/X11/KDE, where when quitting the
game the window switches to the default size (ie 320x200 with 2x zoom) and then
instantly back to whatever the window size was in-game. I documented this
bug in gfx_sdl.bas.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10354 7d344553-34f0-0310-a9b1-970ce8f1c3a2