diriswriteable modified its argument!
It replaced "" with CURDIR.
This bug has existed for a year, since r8559. I don't know whether fixing this
bug will break anything, but any such bug-dependent code would have to have been
introduced in the last year. And there's not much code that distinguishes
between "" and CURDIR anyway, so we're probably OK.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10382 7d344553-34f0-0310-a9b1-970ce8f1c3a2