Behaviour of deletesave+importglobals changed backwards-incompatibly, so fix wrong documentation and tests
Originally (and this was documented), deleting a save didn't delete the global
variables in the save; they could still be read with importglobals. Since
the switch to RSAV in Zenzizenzic (2011), they are deleted.

Using rpgbatch, I found the following games which used both deletesave and
importglobals, so might have been broken by that change:

-Devs Like Waffles
-Baconthulhu
-Paladin Traducer
-Carcere Vicis
-Sprout
-Silhouette
-Chronoboy Adventures
-Perilous Plunge lite
-Super Harem Builder

I checked PPL and Chronoboy weren't affected by examining their scripts ... I
think.
All of the other games are from after 2011 (although the original version of Baconthulhu is from 2009)

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10408 7d344553-34f0-0310-a9b1-970ce8f1c3a2