Remove a couple unneeded calls to get_door_on_map in "get door destination map/id"
This calls replaced original calls to valid_door() before the ability to read
doors on other maps was added. Originally, asking about a non-existent door
would throw an error, but get_door_on_map() only threw an error for a bad map
id.  But these commands have always been documented to return -1 for an invalid
door so I think it's better not to throw an error (except on bad map idea)

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10424 7d344553-34f0-0310-a9b1-970ce8f1c3a2