Add some comments about FB bug 885 "gfxlib may deadlock due to incorrect lock ordering"
https://sourceforge.net/p/fbc/bugs/885/
I just discovered this bug and realised it is responsible for the freezing in
gfx_fb which I fixed in r708 by locking gfxmutex before calling gfx backend
functions.
So in future if this gets fixes, we could theoretically remove most of the
uses of gfxmutex... I'm not sure whether it's a good idea though.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10427 7d344553-34f0-0310-a9b1-970ce8f1c3a2