Fix (apparently) keys getting stuck when using Ctrl-F8 to switch back to gfx_sdl
I don't know whether this is a real fix, because I don't know why it happens
(something internal in SDL), but the problem went away.
It happened 80% of the time for me on both Linux/X11/KDE and Win XP, and Gaplan
also saw it on Windows.

Also, don't call gfx_close twice on switching backends. That was accidental.
Sadly, this doesn't seem to fix anything.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10452 7d344553-34f0-0310-a9b1-970ce8f1c3a2