Allow "seed random" while recording/replaying input. Rename mersenne_twister to reseed_prng and move to util.bas
Disallowing "seed random" will break games which rely on it while recording,
allowing it will break games while replaying. The correct way to get
reproduceable results would be to record the call in the .ohrkeys file - which
we don't do yet. So this is just half of a fix.

Also, improve quality of seeding. Before, calling "seed random" twice on the
same second would give the same result!

And silence mersenne_twistwe debug-spam.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10465 7d344553-34f0-0310-a9b1-970ce8f1c3a2