Don't fade in the screen when calling the on-death script, unless needed
If you died in battle, a fade in is already queued for the next tick anyway.
If you died for some other reason, the screen probably isn't faded out, so
the fadein would cause an unwanted pause.
If the screen really is faded out, fade it in.

Also make sure fadestate is initialised correctly - but it's not actually
reliable

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10484 7d344553-34f0-0310-a9b1-970ce8f1c3a2