Whoops, lodepng.bi was very broken in 64 bit builds - all size_t's were 4 byte instead of 8 byte!
This happened because I translated the header with fbfrog which converts C
size_t to FB uinteger, forgetting that our integer/uinteger is different from
the norm in 64 bit builds.

I'm not aware of whether this actually broke 64 bit builds or not.
But fbc only raised warnings in the code for writing a PNG, which isn't
currently used, so probably not.

I checked all other lib/*.bi headers are OK

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10499 7d344553-34f0-0310-a9b1-970ce8f1c3a2