Fix another bug where pushable NPCs are obstructionless while being pushed against a wall.
Reported by sheamkennedy here: https://www.slimesalad.com/forum/viewtopic.php?t=7572

The original bug is bug 330 "Pushed npcs are obstructionless", fixed in
Voxhumana. That bugfix only fixed the case that an NPC is being pushed against
another NPC, not other obstructions.
Also, the code didn't check for "suspend npc walls" or NPCInst.ignore_walls
either, but that's very insignificant.
Fixed all that and simplified by calling npc_collision_check instead.

This bugfix is only enabled if the existing "Simulate Pushable NPC obstruction bug"
backcompat bit is off. That means that this is technically a backcompat break.
That backcompat bit was added because some games make use of the ability to
push one NPC through another (eg Chase You NPCs on the other side of a pushable
block can catch you if you push the block through them).
Unlike pushing-against-an-NPC, I believe it's safe to change the behaviour
of pushing-against-a-wall because I don't see any use for it, or remember
ever seeing it. Plus the backcompat bit protects the majority of games anyway.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10513 7d344553-34f0-0310-a9b1-970ce8f1c3a2