Remove non-threadsafe filenames map from filelayer.cpp, instead store the names in openfiles[]
filelayer.cpp is now almost threadsafe except for dump_openfiles()

Also, even if a OPENFILE-opened file isn't fully hooked, always hook CLOSE.  The
effect of this is that get_filename() now reliably either returns the filename
(if it was opened with OPENFILE) or says that it doesn't known it. Previously it
could return a stale filename value if it wasn't a hooked file.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10564 7d344553-34f0-0310-a9b1-970ce8f1c3a2