Make sure that stunned combatants can't attack; split out atkq_attack_active(), and
If there was an attack on the attack queue with .delay <= 0 but the attacker was
stunned, it would happen. I'm not sure when that might happen, but probably if
there were multiple queued attacks ready to start, one of which stuns one of the
other attackers. Could happen much easier now that delays can be negative

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10629 7d344553-34f0-0310-a9b1-970ce8f1c3a2