Fix atkq problems in turn-based mode when an attack has a negative delay
Active-time battles not affected

This fixes a bug where attack delays were counted in attacks AND ticks:
a delay of 180 attacks would cause battle to pause for 10 seconds!

Also, if multiple attacks had a negative or zero delay, it would pick the first
one instead of the one with the lowest delay.

Also, if there was an attack with zero delay (which is normally the case for the
first attack in a round!) it would start even if one of the conditions in
turn_mode_time_passage tried to prevent the passage of time. I don't know
whather that was a problem or whether it indicates that those conditions didn't
actually do anything!

Instead of counting down delays by 1 each tick, delays of all attacks get
incremented or decremented so the next attack has delay 0.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10631 7d344553-34f0-0310-a9b1-970ce8f1c3a2