Rewrite plank menu arrow controls (plank_menu_move_cursor) to fix horrible problems
There's a bunch of commented out code here, because this was never really
finished. And some of it is a lot more complicated than it needs to be.
Checking in work I did long ago.

And this doesn't solve all the problems. I think the approach is not ideal.
It would be better to remember the initially selected plank at the start of
a sequence of key strokes, and move in a straight line from that.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10687 7d344553-34f0-0310-a9b1-970ce8f1c3a2