Restore thinggrabber & ability to jump straight to an editor or create new record
Hold down Shift or Ctrl with Enter/Space/Click to go straight to the editor
instead of the browser. This is far more convenient than Ctrl-E to edit.
Updated the hints at the bottom of menus for this new key combination.

Insert/+ will create a new record; it incorrectly also went to the browser.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10695 7d344553-34f0-0310-a9b1-970ce8f1c3a2