One-time upgrade: scan for broken textbox After chains left over from a bug fixed in r9931.
This bug in Dwimmercrafty and Etheldreme happened when you used Ctrl+Left/Right
to link textboxes.

In upgrade() check for textboxes with After chain set but After conditional ==
Never, and show a list of affected textboxes. Once all of them are fixed or the
user asks not to be prompted again a fixbit is set and the check isn't done
again.

I checked a number of large games and couldn't find any that had such a Never
textbox chain, so annoying false positives seem to be rare.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10709 7d344553-34f0-0310-a9b1-970ce8f1c3a2