Pass strings as zstring ptrs to various functions to reduce compiled code size
Total size reduction (gengcc=1 debug=0) is:
ohrrpgce-custom: 168KB
ohrrpgce-game: 115KB
And a bit more when compiled with gengcc=0.
Pretty happy with that!

Passing a string constant as a zstring ptr is a single instruction, while
passing it as a string requires constructing and deleting an FBSTRING, which
add a lot of extra code. If the string is an expression, there's no difference.

I picked out functions which are called a lot, normally with string constants:
-debug/showerror/pop_warning/etc family
-a_append for string arrays
-RELOAD node names & paths
-config read/write functions
-readglobalstring
-multichoice/yesno/etc family
-show_help
-menu item appending functions
-etc

The main thing to watch out for when converting a function to take a zstring ptr
instead of a string is to ensure that len(s) is changed to len(*zs). len(zs)
returns sizeof(zstring ptr)!!

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10726 7d344553-34f0-0310-a9b1-970ce8f1c3a2