Add OPENFILE OR_ERROR flag
The nice thing about using this is that it reports the errno message which
is more informative than just "Couldn't open file".
So always prefer to use OR_ERROR to manually reporting errors.
(If you don't use OR_ERROR and the reason the file couldn't be opened isn't
simply that it's missing, OPENFILE already prints a message anyway)

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10730 7d344553-34f0-0310-a9b1-970ce8f1c3a2