New setkeys-based joystick state tracking. Add JoystickState, JoyScancode
This is a replacement for SUB control()-based polling the joystick,
which has crappy detection of new button presses and key repeat which behaves
differently from setkeys/keyval. These differences are the cause of
joysticks not working correctly in various menus.

This new code treats keyboard buttons and joystick buttons the same,
also getting us most of the way to recording and replay of joystick input.

There's no way yet to actually read this state, or to support a variable
number of joysticks

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10778 7d344553-34f0-0310-a9b1-970ce8f1c3a2