Re-do fix for gfx_sdl (now gfx_sdl2 too) keys getting stuck after backend reinit
I fixed this problem recently in r10452, but it turns out that fix was bad, it
only worked because gfx_backend_menu() calls multichoice() which calls
push_and_reset_gfxio_state() which causes gfx_sdl to recreate the window,
causing keys to be down.
Fixed it in a simpler way.

Also moved gfx_close to gfx_sdl_common.bas

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10821 7d344553-34f0-0310-a9b1-970ce8f1c3a2