Fix memory leaks in ClearHeroData and load_hero_from_reload
The second being a fix for the "Leaked # global reload nodes" warnings
for 'hero' nodes when quitting Custom.

None of these leaks caused problems; they aren't related to the hero data
corrupt bugs fixed in the next commits.

In general we should probably just switch to adding constructors if necessary
for default (although generally the default values can be put straight in the
UDT) and calling destructor+constructor instead of writing "Clear" functions.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10843 7d344553-34f0-0310-a9b1-970ce8f1c3a2