Shop stock levels are reset or adjusted when the shop thing is edited
Changing thing type/id resets the stock instead of using garbage data.  If only
the stock level changed, the stock is increased/decreased by the change in the
initial stock level if possible.

This is accomplished by storing the type, id and initial stock of each shop
thing in gam.original_stock(), and in RSAV.

Also, I added a F6 debug key to shop thing editor to show stockidx.
And ohrhelp/test_game.txt was a bit out of date

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10868 7d344553-34f0-0310-a9b1-970ce8f1c3a2