Fix sprite_empty_cache ignoring its arguments and always emptying whole cache!
This has been broken (in fact, not implemented) ever since it was written in r3213.

(Also, it wouldn't have worked even if it had been implemented, because
SpriteCacheEntry.hash was unsigned, causing a broken signed-unsigned comparison)

The only places that actually called sprite_empty_cache with arguments
were when deleting a frame or importing a sprite from the (new) spriteset
browser.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10909 7d344553-34f0-0310-a9b1-970ce8f1c3a2