Start on game translation support.
A translation is a mapping from codes (like 'tb4' for text box 4) to strings.

Game strings may be exported as a .txt from the SPAM menu, translated by hand or
using a online translator, then import into a game from the debug menu.

The format of the .txt and the parser is designed to tolerate horrible mangling
by an automated translator. Unfortunately they all mangle differently, and this
even depends on the language. Bing Translate is the worst.

Currently, only textboxes translations are implemented, and they're incomplete,
because auto-wrapping and un-wrapping text around portraits isn't implemented,
and neither is preservation of indentation.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10917 7d344553-34f0-0310-a9b1-970ce8f1c3a2