Synchronous HTTP client (GET/POST) implementation using Winsock
Unix support missing.
Multipart POST (e.g. for forms and file upload) also not implemented yet, but
aside from that, it works. But it needs to be converted to asynchronous.

I put it in a new file, os_sockets.c, because it's OS-specific, but will be
shared between Windows and Unix

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10938 7d344553-34f0-0310-a9b1-970ce8f1c3a2