Fix very minor inconsistency in "hero is walking"
When the leader is pathfinding and caterpillar is enabled, heroiswalking
for trailing heroes would return false for one tick at the end of each step.
Now it returns true continually for them, like it does for the leader.

Also mention the "hero/npc is walking" fix for pathfinding in whatsnew,
since James forgot it

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10991 7d344553-34f0-0310-a9b1-970ce8f1c3a2