percent_grabber now works correctly without needing to pass in repr - stores state in a global
Previously, you couldn't enter decimal places unless you correctly stored
the repr string, and passed it in, but that's very inconvenient to do in various
menus.
I avoided adding a global variable for this before, but it's hard to avoid.
This is very similar to the negative_zero global for intgrabber.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@10996 7d344553-34f0-0310-a9b1-970ce8f1c3a2