GURGU.mid testcase for the infamous MIDI crash (bug 2000), from Timestream Saga
This freezes immediately 100% of the time for me, both music_sdl and music_sdl2
on Win XP.

Just appears to be silent when played with SDL_mixer or adlmidi on Linux, but
with timidity or on Windows it plays one note.

Also included the bam file it was translated from.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11217 7d344553-34f0-0310-a9b1-970ce8f1c3a2