bam2mid: Fix a couple of bugs in handling loop points
Both ordinary and return-from-chorus loop points had an off-by-one error in the
file position they jumped to! This meant that at least one command was missed,
with possibly more read as garbage until the stream re-synced.

Also, stop the song if an infinite loop point is encountered. I found one song
(ARIEDUET.BAM) where it happened before the end of the file.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11221 7d344553-34f0-0310-a9b1-970ce8f1c3a2