As an alternative to marking functions PRIVATE you can mark them LOCAL, which does nothing
Marking a function PRIVATE indicates that the function is internal and
shouldn't/can't be called from outside the file. But it also tells the compiler
not to export that symbol and can result in .pdb debug symbols not being
produced when using cv2pdb.  It also encourages the compiler to inline the
function.  So use LOCAL instead of PRIVATE to indicate the same thing (as
documentation only) without debug symbol problems.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11262 7d344553-34f0-0310-a9b1-970ce8f1c3a2