Fix stack corruption causing a possible crash on Windows in get_process_path, os_get_documents_dir
Introduced in r10950 and r1120 (2019-01-30 and 2019-04-10)

Again, several function pointers were declared wrong.

Whether this mistake actually corrupted memory depends on whether -gen gcc was
used or not and compiler decisions. GCC handles the stack differently and seems
to be more resistant to the esp pointer being modified.
In a debug (non-gen gcc) build, I found today that Custom always immediately
crashed inside get_process_path due to a crashed previous instance of Custom
causing the line
  debuginfo "pid " & pid & " image " & ret & " may have crashed, exitcode " & exitcode
to execute. But it didn't crash with -gen gcc, and has never crashed on me in
the last 7 months.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11270 7d344553-34f0-0310-a9b1-970ce8f1c3a2