If genMaxInventory is out of range make sure we don't crash, clamp it.
I don't know whether anyone has experienced a crash this reason, but it could
happen in a corrupt .rpg.

The bug in editing genMaxInventory introduced last year and fixed in r11089
earlier this year could result in a genMaxInventory that wasn't a multiple of 3,
but not a crash-causing out-of-range value.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11273 7d344553-34f0-0310-a9b1-970ce8f1c3a2