Fix pdb=1 debug>=0 builds not containing reference to .pdb file, by stripping symbols
Crash reports from Windows debug builds were still useless even after building
with scons pdb=1.

cv2pdb strips DRAWF debug info at the same time that it adds a section to an
.exe with the .pdb name and GUID, without which the pdb can't be used.  I think
it's been years since we asked anyone to download gdb and produce a backtrace
anyway, but if we did want to, we would now need a separate debug build. I think
it would be possible to modify cv2pdb to keep the DWARF info.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11278 7d344553-34f0-0310-a9b1-970ce8f1c3a2