Err... The "Shop is empty" message wasn't visible
James added that global text string in 2013 when converting Shops to slices, but
I hadn't noticed it existed. Recently in r10926 I made shops autoquit when
they're empty, to restore the pre-2013 behaviour, because Nathan complained it
broke certain uses of shops as menus.

The "Shop is empty" message was hidden behind the "Bought..." alert, so rather
than get rid of it, I moved it up to make it visible for the half a second until
the shop autoquits.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11310 7d344553-34f0-0310-a9b1-970ce8f1c3a2