android: declare/document two other SDL_ANDROID text input functions
I managed to use SDL_ANDROID_ToggleScreenKeyboardTextInput for editing existing
text by adding a previous_text arg to show_virtual_keyboard and calling it from
strgrabber, but it was just so crappy, all kinds of problems and awful user
experience and workarounds needed that it's not worth it. Definitely go with the
slice-based touch_virtual_keyboard instead.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11370 7d344553-34f0-0310-a9b1-970ce8f1c3a2