ohrhelp files weren't found when installed system-wide on Linux
I broke this in r10708 (a year ago) when removing the DATAFILES define,
which also broke searching for data/, but I fixed that problem in r10885.

Note that when installed in a read-only location, the cwd is changed from
exepath, but my change to get_help_dir assumed they were identical.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11372 7d344553-34f0-0310-a9b1-970ce8f1c3a2