mapedit: improved display of multiple NPCs on the same tile
Previously the ID and copy number of the topmost NPC was shown (because that
NPC was drawn on top of the ID/copy of lower NPCs), but since the last commit
all the text was on top of all NPCs. Now, it displays the number of NPCs on the
tile instead, and "Enter: view" at the bottom of the screen.

Also, in the NPC-on-this-tile viewer, display the copy number for each NPC.

Getting some use out of HashTable here.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11397 7d344553-34f0-0310-a9b1-970ce8f1c3a2