Progressively generated minimaps. Rewrote MapPreviewer to use them
MinimapGenerator.run() takes number of seconds to generate for.

MapPreviewer (currently used by the map picker and 'Add map as copy of') has a
bunch of logic to decide whether to delay generation or start by rendering just
a small minimap in order to attempt to:
-not cause a framerate drop even on slow machines
-not cause seizures as you scroll through maps

The mapedit.instant_minimap_threshold config option is no longer used.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11442 7d344553-34f0-0310-a9b1-970ce8f1c3a2