Adjustable ticks per walking animation frame. Defaults to 110ms/frame instead of 2 ticks
That is, existing games which run at >18fps now animate the walking speed at
(approx) the same rate as 18fps games.

All walkabouts/NPCs/heros share the animation rate, including for walk-in-place.
This is just a temporary setting and will be removed again (upgraded) when the
animation editor is available.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11455 7d344553-34f0-0310-a9b1-970ce8f1c3a2